Moved integration tests up one level
